% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongodbtools_functions.R
\name{mdb.findVarsToCSV}
\alias{mdb.findVarsToCSV}
\title{mdb.findVarsToCSV}
\usage{
mdb.findVarsToCSV(rmongodb, strCollection, strQuery, listVars, strFile,
  strDatabase = NULL)

mdb.findVarsToCSV(rmongodb, strCollection, strQuery, listVars, strFile,
  strDatabase = NULL)
}

