% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongolite.R
\name{mdb.mongolite}
\alias{mdb.mongolite}
\title{mdb.mongolite}
\usage{
mdb.mongolite(ip, port, database, collection, user = connData$IAM_user,
  pass = connData$IAM_pass, use_ssl = TRUE,
  use_weak_cert_validation = TRUE)
}

